finish submit application page
