+++ src/components/Navbar.js
@@ -60,16 +60,12 @@ const Navbar = class extends React.Component {
             className={`navbar-menu ${this.state.navBarActiveClass}`}
           >
             <div className="navbar-start has-text-centered">
-              <Link className="navbar-item" to="/about">
-                About
-              </Link>
-              
              
-              <Link className="navbar-item" to="/contact">
-                Contact
+              <Link className="navbar-item" to="/contact/application-download">
+                Download Application
               </Link>
-              <Link className="navbar-item" to="/contact/examples">
-                Form Examples
+              <Link className="navbar-item" to="/contact">
+                Submit Application
               </Link>
             </div>
             <div className="navbar-end has-text-centered">
